AI Navmesh and navigation
Added AI Navmesh and Navigation with 3 AI State implemented for now.

Idle: the AI will reach this location then wait bettween 4 to 15 seconds before changing state.
Patrol: The AI will reach the area and then will instantly change to the next state
Chase: AI will chase the player just need to implement the logic of when he should do that.

If you want to add an AI just add the AiNavigationScript to it. And then just add the "state" and "position" to the list. It will automatically go to the next item on the list once it reach the first destination.