Combat
Added so the player can shoot and kill enemies. Also added a few script to show the equipped weapon transfer the damage from the scriptable object to the bullet. Pickable object can now be of type weapon and has a category and type with a damage it does and other things